LayoutTests/platform/ios-simulator/editing/spelling/spellcheck-async-mutation-expected.txt

CONSOLE MESSAGE: line 211: InvalidAccessError: DOM Exception 15: A parameter or an operation was not supported by the underlying object.
FAIL: Timed out waiting for notifyDone to be called
Test for asynchronous spellchecking in case DOM mutation happens. This test checks crash won't happen if DOM mutations happened.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test Start: 1
FAIL requestId should be >= lastRequestId + 1. Was 0 (of type number).
PASS Request has been processed.

Test Start: 2
FAIL requestId should be >= lastRequestId + 1. Was 0 (of type number).
FAIL Request has been processed but we detected unexpected marker location.

Test Start: 3
FAIL requestId should be >= lastRequestId + 1. Was 0 (of type number).
PASS Request has been processed.

Test Start: 4
FAIL requestId should be >= lastRequestId + 1. Was 0 (of type number).
PASS Request has been processed.

Test Start: 5
FAIL requestId should be >= lastRequestId + 1. Was 0 (of type number).
zz zz zz
